+++ CHANGELOG.md
@@ -35,6 +35,8 @@
     coverage.
 - Changes to SqlServerMaxDop
   - Added en-US localization ([issue #616](https://github.com/PowerShell/SqlServerDsc/issues/616)).
+- Changes to SqlRS
+  - Reporting Services are restarted after changing settings, unless `$SuppressRestart` parameter is set ([issue #1331](https://github.com/PowerShell/SqlServerDsc/issues/1331)). `$SuppressRestart` will also prevent Reporting Services restart after initialization.
 
 ## 12.4.0.0